Fixes #5194: Knob: Move hidden cursor when turning
Before this commit, the cursor was always reset over the knob. Now, it
runs over the screen and tries to flip on the opposite screen side if it
hits the screen border.

This changes the meaning of `Knob::m_origMousePos` from "always being
the position where the user clicked" to "the last position in
mouseMoveEvent".